+++ src/xkblas.cc
@@ -127,9 +127,9 @@ xkblas_t::memory_coherent_async(
     device_global_id_t device_global_id,
     matrix_storage_t storage,
     void * ptr,
-    size_t ld,
-    size_t m,
-    size_t n,
+    int ld,
+    int m,
+    int n,
     size_t sizeof_type
 ) {
     return this->runtime.memory_coherent_async(device_global_id, storage, ptr, ld, m, n, sizeof_type);
@@ -149,9 +149,9 @@ xkblas_t::memory_coherent_sync(
     device_global_id_t device_global_id,
     matrix_storage_t storage,
     void * ptr,
-    size_t ld,
-    size_t m,
-    size_t n,
+    int ld,
+    int m,
+    int n,
     size_t sizeof_type
 ) {
     return this->runtime.memory_coherent_sync(device_global_id, storage, ptr, ld, m, n, sizeof_type);
